rename video file
